requirements.txt
# CrewAI Agent Template Dependencies
# Core framework for orchestrating AI agents
crewai>=0.130.0

# LangChain for tool integration
langchain>=0.1.0

# YAML configuration support
pyyaml>=6.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Additional tools and integrations
# Uncomment the lines below if you need these features

# Web search capabilities
# duckduckgo-search>=4.1.0

# Database support
# sqlalchemy>=2.0.0

# File processing
# python-multipart>=0.0.6

# HTTP requests
# requests>=2.31.0

# Data analysis
# pandas>=2.0.0
# numpy>=1.24.0

# Visualization (optional)
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Google Sheets API integration
gspread>=5.12.4
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Web scraping dependencies
beautifulsoup4
lxml
urllib3

# Additional dependencies
google-api-python-client
openpyxl
requests 